final changes(100% working)

+++ Mappings/Dto's/ReadingDto.cs
@@ -6,21 +6,21 @@ namespace WaterBillingMobileAppAPi.Mappings.Dto_s
 
     {
         [PrimaryKey, AutoIncrement]
-        public long WaterReadingExportDataID { get; set; }
+        public int WaterReadingExportDataID { get; set; }
         public string CUSTOMER_NUMBER { get; set; }
         public string? METER_NUMBER { get; set; }
         public decimal CURRENT_READING { get; set; }
         public decimal PREVIOUS_READING { get; set; }
-        public long MonthID { get; set; }
-        public long Year { get; set; }
-        public long WaterReadingExportID { get; set; }
+        public int MonthID { get; set; }
+        public int Year { get; set; }
+        public int WaterReadingExportID { get; set; }
 
     }
 
     public class UpdateReadingDto
 
     {
-        public long WaterReadingExportDataID { get; set; }
+        public System.Int64 WaterReadingExportDataID { get; set; }
         public decimal CURRENT_READING { get; set; }
         public string Comment { get; set; }
         public string? METER_READER { get; set; }
